refactor: App.js 로직 Baseball.js로 이동. 컴포넌트 분리
